Use the responses.legislative_period_id for lookup
Previously it was incorrectly using the legislative_period_id from the
legislative_periods table, which is the EveryPolitician legislative
period, but actually we wanted it to be using the legislative_periods.id
field which in the responses table is called legislative_period_id.

This switches to selecting just the
'responses.legislative_period_id' field and then using that to lookup
the LegislativePeriod model by its primary key.