Fix error running rake tasks and migrations
The 'rake cache:country_person_counts' task needs to be run before
migration 006 is run so that the CountryCount model is correctly
populated. However at that point the CountryCount models doesn't have
the 'gender_count' column because that is added in a later migration.
This checks that the CountryCount model responds to 'gender_count='
before assigning the gender_count.