Cache counts after migrating db during provisioning
This will only work if provisioning a fresh box, if you've already
provisioned a box then this will probably fail because the migrations
need to be run up to a certain point and then one of the cache tasks
needs to be run, then more migrations followed by more rake tasks.

The easiest way to avoid this pain when working with vagrant is to
'vagrant destroy' your box and then do a fresh 'vagrant up'.