Don't render countries with no person/gender counts
In the main list of countries we don't want to show countries where we
haven't cached the counts for people and number of people that already
have gender.

Once the counts have been cached (which should happen within a minute or
two of a country being added) then the country will display correctly.