Only preload the first two images for a game
For large terms preloading all of the images slows things down
significantly. So instead of preloading them all we just preload the
first two visible images. The rest of the images have the 'src'
attribute in a 'data-src' attribute, which then gets copied across when
it's added to the back of the stack.