Save onboarding completion server side
Previously we were marking the user as having completed onboarding by
triggering an AJAX request at the end of the onboarding process. However
this potentially caused issues when the user was on a slow connection as
they could finish onboarding and click the link at the end before the
AJAX request had completed.

This change makes the link at the end of onboarding to go to
/onboarding-complete, which then sets the flag on the user or the cookie
in the user's browser.