Don't make the user go through onboarding twice
This change means that the "completed_onboarding" session flag always
gets set and then doesn't get removed when the user signs in. This means
that even if the user signs out, when they sign in again they won't be
made to go through onboarding again if they just click play.