Hide level complete message with CSS
Add a special class for level complete messages that need to be hidden
when the page loads. This only applies to the level complete message
that needs to sit behind the stack of person cards. Other level complete
messages such as the one displayed when the country has been completed
don't need this styling applied.