Add LegacyIdMapper class
EveryPolitician is introducing UUIDs globally for all people. This means
that the ids that GenderBalance has stored will no longer map to people
and we need to map new UUIDs back to the legacy ids so we don't loose
scores for past responses.

The legacy ids are still included in the popolo JSON under the
identifiers array with type 'everypolitician_legacy'. This means we can
create a mapping from new UUIDs back to the legacy id.

This class provides that mapping by taking a parsed popolo JSON object
and turning that into a mapping from UUID to legacy ids.