Remove duplicate votes (responses) from GenderBalance
Up until now there was no unique constraint on (user_id, politician_id)
for the responses table. This has resulted in people voting for
politicians gender more than once. This makes it tricky to do things
such as calculate voting patterns and see if there has been a consensus
for a country.