Move methods from helpers into CountryProxy
Rather than have a bunch of functional methods which take a country
as an argument we use the CountryProxy class to keep all these methods
in one place. This makes the views much easier to follow and change
since method calls take the form 'country.complete?' rather than
'complete?(country)'.