Move Download data button to legislature level and wire up
We currently only offer exports at the legislature level, so move the
button to reflect this.

Note: This may be confusing to users since the data they download from
Gender Balance might well be different from the data that's displayed on
the report page, since that data is coming from EveryPolitician.